pyproject.toml
[project]
name = "pywebview-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pywebview>=6.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
